llvm-stress testcase generation

+++ run_linux.sh
@@ -1,3 +1,13 @@
+if [ -z "$1" ]; then
+    echo "Usage: $0 <input_file>"
+    exit 1
+fi
+
+TEST_DIR=test/llvm-stress
+INPUT=$1
+OUTPUT=optimized_$INPUT
+echo "INPUT: " $1
+
 rm -rf build
 mkdir build
 
@@ -11,15 +21,14 @@ clang++ -g -S -emit-llvm ml55822_mp46753-loop-opt-pass.cpp $(llvm-config --cxxfl
 
 mv *.ll ll/
 
-cd test
-clang++ -g -S -emit-llvm *.cpp $(llvm-config --cxxflags)
-mv *.ll ./ll/
-cd ../
+#cd test
+#clang++ -g -S -emit-llvm *.cpp $(llvm-config --cxxflags)
+#mv *.ll ./ll/
+#cd ../
 
-TEST_DIR=test/ll
-INPUT=test1.ll
 #TEST_DIR=ll
 #INPUT=ml55822_mp46753-loop-analysis-pass.ll
 
 opt -load-pass-plugin ./build/libloop-analysis-pass.so -passes="ml55822_mp46753-loop-analysis-pass" <"$TEST_DIR/$INPUT" >/dev/null 2> log/analysis_output.log
-opt -load-pass-plugin ./build/libloop-opt-pass.so -passes="ml55822_mp46753-loop-opt-pass" <"$TEST_DIR/$INPUT" >/dev/null 2> log/opt_output.log
+#opt -load-pass-plugin ./build/libloop-opt-pass.so -passes="ml55822_mp46753-loop-opt-pass" -S -o $TEST_DIR/$OUTPUT <"$TEST_DIR/$INPUT" >/dev/null 2> log/opt_output.log
+opt -load-pass-plugin ./build/libloop-opt-pass.so -passes="ml55822_mp46753-loop-opt-pass" $TEST_DIR/$INPUT -S -o $TEST_DIR/$OUTPUT >/dev/null 2> log/opt_output.log
